Added Bulk Documentation and Redesigned workflows and packaging files.
