+++ .github/workflows/python-publish.yml
@@ -26,7 +26,9 @@ jobs:
         pip install setuptools wheel twine
     - name: Build and publish
       env:
-        TWINE_USERNAME: ${{ secrets.PYPI_USERNAME }}
-        TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
+        USER: ${{ secrets.PYPI_USERNAME }}
+        PASSW: ${{ secrets.PYPI_PASSWORD }}
       run: |
-        python setup.py sdist bdist_wheel
+        python setup.py sdist
+        twine upload dist/* -u $USER -p $PASSW
+        rm -rf build/ dist/ *.egg-info/
