+++ discordgame/errors.py
@@ -1,0 +1,2 @@
+class GameError(Exception):
+    pass
